Support AOT-based autowiring for test class instances
This commit introduces support for AOT-based autowiring for test class
instances via a new, auto-registered
AotDependencyInjectionTestExecutionListener that uses the
InjectionPointsSupplier from the spring-aot module.

The InjectionPointsSupplier had to be made public for this approach.

This approach effectively copies a subset of the functionality of the
standard AutowiredAnnotationBeanPostProcessor in order to autowire
methods and fields in the test instance based on the information
obtained from the InjectionPointsSupplier.

See gh-1169